Refactors route adding a controller
