+++ api/routes/pokemonRoutes.js
@@ -1,28 +1,7 @@
 const express = require('express');
 const router = express.Router();
-const { getPokemonData } = require('../integration/pokeapiIntegration');
-const { processPokemonData } = require('../services/pokemonService');
-const PokemonNotFoundException = require("../exceptions/PokemonNotFoundException");
-const PokemonDataFetchException = require("../exceptions/PokemonDataFetchException");
+const { getPokemonByName } = require('../controllers/pokemonController'); // Import the controller function
 
-router.get('/pokemon/:name', async (req, res) => {
-    const pokemonName = req.params.name;
-    try {
-        const pokemonName = req.params.name;
-        const rawData = await getPokemonData(pokemonName);
-        const processedData = processPokemonData(rawData);
-        res.json(processedData);
-    } catch (error) {
-        if (error instanceof PokemonNotFoundException) {
-            return res.status(error.status).json({ error: error.message });
-        }
-
-        if (error instanceof PokemonDataFetchException) {
-            return res.status(error.status).json({ error: error.message });
-        }
-
-        res.status(500).json({ error: 'An unexpected error occurred' });
-    }
-});
+router.get('/pokemon/:name', getPokemonByName); // Use the controller function
 
 module.exports = router;
